+++ app/src/main/java/jp/atsushitechacademy/kanamori/calcapp/SecondActivity.java
@@ -14,7 +14,7 @@ public class SecondActivity extends AppCompatActivity  {
         setContentView(R.layout.activity_second);
 
         Intent intent = getIntent();
-        int value = intent.getIntExtra("VALUE", 0);
+        double value = intent.getDoubleExtra("VALUE", 0.0);
 
         TextView textView = (TextView) findViewById(R.id.textView);
         textView.setText(String.valueOf(value));
